// AngularShopFrontend/src/app/featuers/profile/models/address.ts
export interface Address {
  id: number;
  isMain: boolean | any;
  state: string;
  city: string;
  firstName: string;
  fullAddress: string;
  lastName: string;
  number: string;
  postalCode: string;
}
export interface ShipToAddress {
  isMain: boolean | any;
  state: string;
  city: string;
  firstName: string;
  fullAddress: string;
  lastName: string;
  number: string;
  postalCode: string;
}